+++ Firmware/DumbBadgeFirmware/src/console.c
@@ -12,20 +12,21 @@
 #include "LCD.h"
 #include "console.h"
 #include "fonts.h"
+
 #include "globals.h"
 
-uint16_t xCharPos = 0;
-uint16_t yCharPos = 0;
+uint8_t xCharPos = 0;
+uint8_t yCharPos = 0;
 
 char console[80][24];
 
 void drawChar(uint8_t character)
 {
-	uint16_t x = xCharPos;
-	uint16_t y = yCharPos;
+	int x = xCharPos;
+	int y = yCharPos;
 	REG_PORT_OUTCLR1 = LCD_CS;
 	setXY(x*10,y*20,x*10+9,y*20+19);
-	for(uint16_t i=0; i <= 24; i++)
+	for(uint8_t i=0; i <= 24; i++)
 	{
 		for(int j=0;j<8;j++)
 		{
